fix(#369): code review findings — move edge-router to lib/, dedupe ArrowConfig
- Move src/features/editor/edge-router.{ts,test.ts} to src/lib/
- Export ArrowConfig from flow-engine.ts (single source of truth)
- Strengthen unrelated-arrows test: assert routing differs from alone case
